Add history reporting for all unit test (#42)

+++ test/Unit/RunTests.ps1
@@ -4,17 +4,22 @@ if( "" -eq $outputRoot ) { $output = "$PSScriptRoot" } else { $output = "$output
 function DotNet_Test {
   param($testProject)
   $results = "$output\TestResults\" + $testProject.split('.')[0].ToLower()
-  $history = "$PSScriptRoot\$testProject\Testdata\history"
+  $history = "$results\history"
 
-  # Remove existing results
-  if( Test-Path "$results" ) { Remove-Item -Path "$results\*" -Recurse | Out-Null }
-
-  # Build project and run unit tests
-  Set-Location "$PSScriptRoot\$testProject"
+  # Build projects and run unit tests
+  Push-Location "$PSScriptRoot\$testProject"
   if( "$PSScriptRoot" -eq "$output" ) { dotnet clean | Out-Null }
   dotnet build --no-restore
+  if( $LASTEXITCODE -gt 0 ) { Pop-Location; exit 1 }
+  if( Test-Path "$results" ) { Remove-Item -Path "$results\*" -Recurse | Out-Null }
   dotnet test --collect:"XPlat Code Coverage" --no-build --results-directory "$results"
-  Set-Location $PSScriptRoot
+  Pop-Location
+
+  # Copy the oldest 3 history files
+  $wrk = "$PSScriptRoot\$testProject\Testdata\history"
+  if( !(Test-Path "$history") ) { New-Item -ItemType Directory -Path "$history" | Out-Null }
+  foreach( $file in ( Get-ChildItem "$wrk" -Filter *.xml | Sort-Object -Property FullName | Select-Object -First 3 ) )
+  { Copy-Item -Path "$file" -Destination "$history" }
 
   # If running locally generate reports
   if( "$PSScriptRoot" -eq "$output" ) {
@@ -24,24 +29,7 @@ function DotNet_Test {
     $title = "-title:$testProject"
 
     Write-Host "Generating coverage reports for $testProject" -ForegroundColor Yellow
-    reportgenerator -reports:"$reports" -targetdir:"$results\reports" $title -verbosity:Warning $rpTypes $setting
-
-     # Tidy up output
-    $wrk = "$results\reports\Html"
-    if( Test-Path "$wrk" ) { Rename-Item -path "$wrk" -NewName "html" }
-
-    $wrk = "$results\reports\Badges" # Move required badges
-    if( Test-Path "$wrk" ) {
-      Remove-Item "$wrk\badge_shieldsio_*.svg" | Out-Null
-      Move-Item -Path "$wrk\*.svg" -Destination "$results\reports"
-      Remove-Item "$wrk"
-    }
-
-    $wrk = "$results\reports\MarkdownSummaryGithub" # Move required markdown
-    if( Test-Path "$wrk" ) {
-      Move-Item -Path "$wrk\*.md" -Destination "$results\reports"
-      Remove-Item "$wrk"
-    }
+    reportgenerator -reports:"$reports" -targetdir:"$results\reports" -historydir:$history $title -verbosity:Warning $rpTypes $setting
   }
 }
